+++ CHANGELOG.md
@@ -23,7 +23,8 @@ use [Keep a Changelog] (http://keepachangelog.com/).
 
 - Normalize headers in options
 - Normalize default headers
-- Non-utf8 header value
+- Non-utf8 header value for HAR generation
+- Non-utf8 querystring value for HAR generation
 
 ## [2.0.0-beta5] - 2022-09-07